requirements.txt
Django>=3.0,<=4.2
psycopg2>=2.8
black
mypy
pylint
djangorestframework
pylint-django